// mcp_servers/resources/oai-apps/worker-profile/src/types.ts
export interface WorkerProfile {
  workdayId: string;
  workerId: string;
  name: string;
  email: string;
  workerType: string;
  businessTitle: string;
  location: string;
  locationId: string;
  country: string;
  countryCode: string;
  supervisoryOrganization: string;
  jobType: string;
  jobProfile: string;
  primaryJobId: string;
  primaryJobDescriptor: string;
}